# docs/site/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Memobase API",
        "summary": "APIs for Memobase, a user memory system for LLM Apps",
        "version": "0.0.17"
    },
    "paths": {
...
        "/api/v1/project/profile_config": {
            "get": {
                "tags": [
                    "project"
                ],
                "summary": "Get Project Profile Config String",
                "operationId": "get_project_profile_config_string_api_v1_project_profile_config_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProfileConfigDataResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "project"
                ],
                "summary": "Update Project Profile Config",
                "operationId": "update_project_profile_config_api_v1_project_profile_config_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProfileConfigData",
                                "description": "The profile config to update"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Create User",
                "description": "Create a new user with additional data",
                "operationId": "create_user_api_v1_users_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserData",
                                "description": "User data for creating a new user"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IdResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/{user_id}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get User",
                "operationId": "get_user_api_v1_users__user_id__get",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "The ID of the user to retrieve",
                            "title": "User Id"
                        },
                        "description": "The ID of the user to retrieve"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserDataResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "Update User",
                "operationId": "update_user_api_v1_users__user_id__put",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "The ID of the user to update",
                            "title": "User Id"
                        },
                        "description": "The ID of the user to update"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "description": "Updated user data",
                                "title": "User Data"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IdResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "user"
                ],
                "summary": "Delete User",
                "operationId": "delete_user_api_v1_users__user_id__delete",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "The ID of the user to delete",
                            "title": "User Id"
                        },
                        "description": "The ID of the user to delete"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/users/blobs/{user_id}/{blob_type}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get User All Blobs",
                "operationId": "get_user_all_blobs_api_v1_users_blobs__user_id___blob_type__get",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "The ID of the user to fetch blobs for",
                            "title": "User Id"
                        },
                        "description": "The ID of the user to fetch blobs for"
                    },
                    {
                        "name": "blob_type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/BlobType",
                            "description": "The type of blobs to retrieve"
                        },
                        "description": "The type of blobs to retrieve"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "description": "Page number for pagination, starting from 0",
                            "default": 0,
                            "title": "Page"
                        },
                        "description": "Page number for pagination, starting from 0"
                    },
                    {
                        "name": "page_size",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "description": "Number of items per page, default is 10",
                            "default": 10,
                            "title": "Page Size"
                        },
                        "description": "Number of items per page, default is 10"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IdsResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/blobs/insert/{user_id}": {
            "post": {
                "tags": [
                    "blob"
                ],
                "summary": "Insert Blob",
                "operationId": "insert_blob_api_v1_blobs_insert__user_id__post",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "The ID of the user to insert the blob for",
                            "title": "User Id"
                        },
                        "description": "The ID of the user to insert the blob for"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BlobData",
                                "description": "The blob data to insert"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/IdResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/v1/users/buffer/{user_id}/{buffer_type}": {
            "post": {
                "tags": [
                    "buffer"
                ],
                "summary": "Flush Buffer",
                "description": "Get the real-time user profiles for long term memory",
                "operationId": "flush_buffer_api_v1_users_buffer__user_id___buffer_type__post",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "description": "The ID of the user",
                            "title": "User Id"
                        },
                        "description": "The ID of the user"
                    },
                    {
                        "name": "buffer_type",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/BlobType",
                            "description": "The type of buffer to flush"
                        },
                        "description": "The type of buffer to flush"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BaseResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BlobType": {
                "type": "string",
                "enum": [
                    "chat",
                    "doc",
                    "image",
                    "code",
                    "transcript"
                ],
                "title": "BlobType"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "ProfileConfigData": {
                "properties": {
                    "profile_config": {
                        "type": "string",
                        "title": "Profile Config",
                        "description": "Profile config string"
                    }
                },
                "type": "object",
                "required": [
                    "profile_config"
                ],
                "title": "ProfileConfigData"
            },
            "ProfileConfigDataResponse": {
                "properties": {
                    "data": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/ProfileConfigData"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "description": "Response containing profile config data"
                    },
                    "errno": {
                        "$ref": "#/components/schemas/CODE",
                        "description": "Error code, 0 means success",
                        "default": 0
                    },
                    "errmsg": {
                        "type": "string",
                        "title": "Errmsg",
                        "description": "Error message, empty when success",
                        "default": ""
                    }
                },
                "type": "object",
                "title": "ProfileConfigDataResponse"
            },
...
        }
    }
}